+++ src/Longbow.Modbus/Interfaces/IModbusMessageBuilder.cs
@@ -1,0 +1,83 @@
+﻿// Copyright (c) Argo Zhang (argo@live.ca). All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+// Website: https://github.com/LongbowExtensions/
+
+namespace Longbow.Modbus;
+
+/// <summary>
+/// Modbus 消息构建器接口
+/// </summary>
+public interface IModbusMessageBuilder
+{
+    /// <summary>
+    /// 构建 Modbus 读取消息方法
+    /// </summary>
+    /// <param name="slaveAddress"></param>
+    /// <param name="functionCode"></param>
+    /// <param name="startAddress"></param>
+    /// <param name="numberOfPoints"></param>
+    /// <returns></returns>
+    ReadOnlyMemory<byte> BuildReadRequest(byte slaveAddress, byte functionCode, ushort startAddress, ushort numberOfPoints);
+
+    /// <summary>
+    /// 构建 Modbus 写入消息方法
+    /// </summary>
+    /// <param name="slaveAddress"></param>
+    /// <param name="functionCode"></param>
+    /// <param name="data"></param>
+    /// <returns></returns>
+    ReadOnlyMemory<byte> BuildWriteRequest(byte slaveAddress, byte functionCode, ReadOnlyMemory<byte> data);
+
+    /// <summary>
+    /// 验证 Modbus 读取响应消息方法
+    /// </summary>
+    /// <param name="response"></param>
+    /// <param name="slaveAddress"></param>
+    /// <param name="functionCode"></param>
+    /// <param name="exception"></param>
+    /// <returns></returns>
+    bool TryValidateReadResponse(ReadOnlyMemory<byte> response, byte slaveAddress, byte functionCode, [NotNullWhen(false)] out Exception? exception);
+
+    /// <summary>
+    /// 验证 Modbus 写入响应消息方法
+    /// </summary>
+    /// <param name="response"></param>
+    /// <param name="slaveAddress"></param>
+    /// <param name="functionCode"></param>
+    /// <param name="data"></param>
+    /// <param name="exception"></param>
+    /// <returns></returns>
+    bool TryValidateWriteResponse(ReadOnlyMemory<byte> response, byte slaveAddress, byte functionCode, ReadOnlyMemory<byte> data, [NotNullWhen(false)] out Exception? exception);
+
+    /// <summary>
+    /// 解析布尔值数组算法
+    /// </summary>
+    /// <param name="response"></param>
+    /// <param name="numberOfPoints"></param>
+    /// <returns></returns>
+    bool[] ReadBoolValues(ReadOnlyMemory<byte> response, ushort numberOfPoints);
+
+    /// <summary>
+    /// 解析无符号短整数数组算法
+    /// </summary>
+    /// <param name="response"></param>
+    /// <param name="numberOfPoints"></param>
+    /// <returns></returns>
+    ushort[] ReadUShortValues(ReadOnlyMemory<byte> response, ushort numberOfPoints);
+
+    /// <summary>
+    /// 将布尔值数组转换为字节数组算法
+    /// </summary>
+    /// <param name="address"></param>
+    /// <param name="values"></param>
+    /// <returns></returns>
+    ReadOnlyMemory<byte> WriteBoolValues(ushort address, bool[] values);
+
+    /// <summary>
+    /// 将无符号短整数数组转换为字节数组算法
+    /// </summary>
+    /// <param name="address"></param>
+    /// <param name="values"></param>
+    /// <returns></returns>
+    ReadOnlyMemory<byte> WriteUShortValues(ushort address, ushort[] values);
+}
