+++ src/Longbow.Modbus/Rtu/DefaultRtuClient.cs
@@ -6,7 +6,7 @@ using System.IO.Ports;
 
 namespace Longbow.Modbus;
 
-class DefaultRtuClient(ModbusRtuClientOptions options, IModbusRtuMessageBuilder builder) : ModbusClientBase, IModbusRtuClient
+class DefaultRtuClient(ModbusRtuClientOptions options, IModbusRtuMessageBuilder builder) : ModbusClientBase(builder), IModbusRtuClient
 {
     private TaskCompletionSource? _readTaskCompletionSource;
     private CancellationTokenSource? _receiveCancellationTokenSource;
@@ -44,7 +44,7 @@ class DefaultRtuClient(ModbusRtuClientOptions options, IModbusRtuMessageBuilder 
         return ret;
     }
 
-    private async Task<ReadOnlyMemory<byte>> SendAsync(ReadOnlyMemory<byte> request)
+    protected override async Task<ReadOnlyMemory<byte>> SendAsync(ReadOnlyMemory<byte> request)
     {
         // 取消等待读取的任务
         _readTaskCompletionSource?.TrySetCanceled();
@@ -64,53 +64,6 @@ class DefaultRtuClient(ModbusRtuClientOptions options, IModbusRtuMessageBuilder 
         await _readTaskCompletionSource.Task.WaitAsync(_receiveCancellationTokenSource.Token);
 
         return _buffer.ToArray();
-    }
-
-    protected override async ValueTask<ReadOnlyMemory<byte>> ReadAsync(byte slaveAddress, byte functionCode, ushort startAddress, ushort numberOfPoints)
-    {
-        // 构建请求报文
-        var request = builder.BuildReadRequest(slaveAddress, functionCode, startAddress, numberOfPoints);
-
-        // 发送请求
-        var received = await SendAsync(request);
-
-        // 验证响应报文
-        var valid = builder.TryValidateReadResponse(received, slaveAddress, functionCode, out var exception);
-
-        Exception = valid ? null : exception;
-        return valid ? received : default;
-    }
-
-    protected override async ValueTask<bool> WriteBoolValuesAsync(byte slaveAddress, byte functionCode, ushort address, bool[] values)
-    {
-        // 构建请求报文
-        var data = builder.WriteBoolValues(address, values);
-        var request = builder.BuildWriteRequest(slaveAddress, functionCode, data);
-
-        // 发送请求
-        var received = await SendAsync(request);
-
-        // 验证响应报文
-        var valid = builder.TryValidateWriteResponse(received, slaveAddress, functionCode, data, out var exception);
-
-        Exception = valid ? null : exception;
-        return valid;
-    }
-
-    protected override async ValueTask<bool> WriteUShortValuesAsync(byte slaveAddress, byte functionCode, ushort address, ushort[] values)
-    {
-        // 构建请求报文
-        var data = builder.WriteUShortValues(address, values);
-        var request = builder.BuildWriteRequest(slaveAddress, functionCode, data);
-
-        // 发送请求
-        var received = await SendAsync(request);
-
-        // 验证响应报文
-        var valid = builder.TryValidateWriteResponse(received, slaveAddress, functionCode, data, out var exception);
-
-        Exception = valid ? null : exception;
-        return valid;
     }
 
     private SerialPort GetSerialPort()
@@ -154,14 +107,10 @@ class DefaultRtuClient(ModbusRtuClientOptions options, IModbusRtuMessageBuilder 
         _readTaskCompletionSource?.TrySetResult();
     }
 
-    protected override bool[] ReadBoolValues(ReadOnlyMemory<byte> response, ushort numberOfPoints) => builder.ReadBoolValues(response, numberOfPoints);
-
-    protected override ushort[] ReadUShortValues(ReadOnlyMemory<byte> response, ushort numberOfPoints) => builder.ReadUShortValues(response, numberOfPoints);
-
     /// <summary>
     /// <inheritdoc/>
     /// </summary>
-    public ValueTask CloseAsync()
+    public override ValueTask CloseAsync()
     {
         // 取消等待读取的任务
         if (_readTaskCompletionSource != null)
@@ -183,18 +132,5 @@ class DefaultRtuClient(ModbusRtuClientOptions options, IModbusRtuMessageBuilder 
             _serialPort.Close();
         }
         return ValueTask.CompletedTask;
-    }
-
-    /// <summary>
-    /// <inheritdoc/>
-    /// </summary>
-    /// <param name="disposing"></param>
-    /// <returns></returns>
-    protected override async ValueTask DisposeAsync(bool disposing)
-    {
-        if (disposing)
-        {
-            await CloseAsync();
-        }
     }
 }
