+++ src/Longbow.Modbus/Tcp/IModbusTcpClient.cs
@@ -20,9 +20,4 @@ public interface IModbusTcpClient : IModbusClient
     /// <param name="token"></param>
     /// <returns></returns>
     ValueTask<bool> ConnectAsync(IPEndPoint endPoint, CancellationToken token = default);
-
-    /// <summary>
-    /// 断开连接方法
-    /// </summary>
-    ValueTask CloseAsync();
 }
