+++ src/Longbow.Modbus/Udp/DefaultUdpClient.cs
@@ -7,7 +7,7 @@ using System.Net.Sockets;
 
 namespace Longbow.Modbus;
 
-class DefaultUdpClient(ModbusUdpClientOptions options, IModbusTcpMessageBuilder builder) : ModbusClientBase, IModbusUdpClient
+class DefaultUdpClient(ModbusUdpClientOptions options, IModbusTcpMessageBuilder builder) : ModbusClientBase(builder), IModbusUdpClient
 {
     private UdpClient _client = default!;
 
@@ -42,7 +42,7 @@ class DefaultUdpClient(ModbusUdpClientOptions options, IModbusTcpMessageBuilder 
         return ret;
     }
 
-    private async Task<ReadOnlyMemory<byte>> SendAsync(ReadOnlyMemory<byte> request)
+    protected override async Task<ReadOnlyMemory<byte>> SendAsync(ReadOnlyMemory<byte> request)
     {
         _client.ThrowIfNotConnected();
 
@@ -64,61 +64,10 @@ class DefaultUdpClient(ModbusUdpClientOptions options, IModbusTcpMessageBuilder 
         return ret;
     }
 
-    protected override async ValueTask<ReadOnlyMemory<byte>> ReadAsync(byte slaveAddress, byte functionCode, ushort startAddress, ushort numberOfPoints)
-    {
-        // 构建请求报文
-        var request = builder.BuildReadRequest(slaveAddress, functionCode, startAddress, numberOfPoints);
-
-        // 发送请求
-        var received = await SendAsync(request);
-
-        // 验证响应报文
-        var valid = builder.TryValidateReadResponse(received, slaveAddress, functionCode, out var exception);
-
-        Exception = valid ? null : exception;
-        return valid ? received : default;
-    }
-
-    protected override bool[] ReadBoolValues(ReadOnlyMemory<byte> response, ushort numberOfPoints) => builder.ReadBoolValues(response, numberOfPoints);
-
-    protected override ushort[] ReadUShortValues(ReadOnlyMemory<byte> response, ushort numberOfPoints) => builder.ReadUShortValues(response, numberOfPoints);
-
-    protected override async ValueTask<bool> WriteBoolValuesAsync(byte slaveAddress, byte functionCode, ushort address, bool[] values)
-    {
-        // 构建请求报文
-        var data = builder.WriteBoolValues(address, values);
-        var request = builder.BuildWriteRequest(slaveAddress, functionCode, data);
-
-        // 发送请求
-        var received = await SendAsync(request);
-
-        // 验证响应报文
-        var valid = builder.TryValidateWriteResponse(received, slaveAddress, functionCode, data, out var exception);
-
-        Exception = valid ? null : exception;
-        return valid;
-    }
-
-    protected override async ValueTask<bool> WriteUShortValuesAsync(byte slaveAddress, byte functionCode, ushort address, ushort[] values)
-    {
-        // 构建请求报文
-        var data = builder.WriteUShortValues(address, values);
-        var request = builder.BuildWriteRequest(slaveAddress, functionCode, data);
-
-        // 发送请求
-        var received = await SendAsync(request);
-
-        // 验证响应报文
-        var valid = builder.TryValidateWriteResponse(received, slaveAddress, functionCode, data, out var exception);
-
-        Exception = valid ? null : exception;
-        return valid;
-    }
-
     /// <summary>
     /// <inheritdoc/>
     /// </summary>
-    public ValueTask CloseAsync()
+    public override ValueTask CloseAsync()
     {
         if (_client != null)
         {
@@ -128,18 +77,5 @@ class DefaultUdpClient(ModbusUdpClientOptions options, IModbusTcpMessageBuilder 
         }
 
         return ValueTask.CompletedTask;
-    }
-
-    /// <summary>
-    /// <inheritdoc/>
-    /// </summary>
-    /// <param name="disposing"></param>
-    /// <returns></returns>
-    protected override async ValueTask DisposeAsync(bool disposing)
-    {
-        if (disposing)
-        {
-            await CloseAsync();
-        }
     }
 }
